BackEnd/requirements.txt
djangorestframework
django-cors-headers
psycopg2
psycopg2-binary

